# server/api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "MTC API Documentation",
    "version": "1.0.0",
    "title": "MTC API Documentation"
  },
  "host": "localhost:10010",
  "basePath": "/api",
  "schemes": [
    "http",
    "https"
  ],
  "parameters": {
    "clientId": {
      "name": "clientId",
      "type": "string",
      "in": "header",
      "required": true,
      "description": "APIs are secured by clientId."
    },
    "clientSecret": {
      "name": "clientSecret",
      "type": "string",
      "in": "header",
      "required": true,
      "description": "clientSecret"
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "paths": {
...
    "/order/invoice/{orderId}": {
      "get": {
        "tags": [
          "Order"
        ],
        "summary": "API for download invoice ",
        "consumes": [
          "application/json"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "operationId": "downloadOrderInvoice",
        "description": "API for download invoice \n",
        "produces": [
          "application/pdf"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "orderId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "file"
            }
          }
        }
      }
    },
    "/order": {
      "parameters": [
        {
          "$ref": "#/parameters/clientId"
        },
        {
          "$ref": "#/parameters/clientSecret"
        }
      ],
      "post": {
        "tags": [
          "Order"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "API for create order",
        "consumes": [
          "application/json"
        ],
        "operationId": "createOrder",
        "description": "API for create Order\n",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "addressId",
                "cart"
              ],
              "properties": {
                "addressId": {
                  "type": "string"
                },
                "cart": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/OrderCart"
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": " Register user successfully ",
            "schema": {
              "type": "object"
            }
          }
        }
      },
      "get": {
        "tags": [
          "Order"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "API for getOrder",
        "consumes": [
          "application/json"
        ],
        "operationId": "getOrder",
        "description": "API for get Order\n",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "offset",
            "minimum": 0,
            "required": false,
            "type": "number"
          },
          {
            "in": "query",
            "name": "id",
            "required": false,
            "type": "string"
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "type": "string"
          },
          {
            "in": "query",
            "name": "limit",
            "minimum": 0,
            "maximum": 20,
            "required": false,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": " get Cart successfully ",
            "schema": {
              "type": "object"
            }
          }
        }
      }
    },
    "/order/{id}": {
      "parameters": [
        {
          "$ref": "#/parameters/clientId"
        },
        {
          "$ref": "#/parameters/clientSecret"
        }
      ],
      "put": {
        "tags": [
          "Order"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "API for update order",
        "consumes": [
          "application/json"
        ],
        "operationId": "updateOrder",
        "description": "API for update Order\n",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "addressId": {
                  "type": "string"
                },
                "cart": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/UpdateOrderCart"
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "update Order successfully ",
            "schema": {
              "type": "object"
            }
          }
        }
      }
    },
    "/order/cancel/{id}": {
      "parameters": [
        {
          "$ref": "#/parameters/clientId"
        },
        {
          "$ref": "#/parameters/clientSecret"
        }
      ],
      "put": {
        "tags": [
          "Order"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "API for cancel order",
        "consumes": [
          "application/json"
        ],
        "operationId": "cancelOrderStatus",
        "description": "API for cancel Order Order\n",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": " Order Cancel successfully ",
            "schema": {
              "type": "object"
            }
          }
        }
      }
    },
    "/order/reorder/{id}": {
      "parameters": [
        {
          "$ref": "#/parameters/clientId"
        },
        {
          "$ref": "#/parameters/clientSecret"
        }
      ],
      "put": {
        "tags": [
          "Order"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "API for re order",
        "consumes": [
          "application/json"
        ],
        "operationId": "reOrder",
        "description": "API for reorder Order\n",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "reOrder successfully ",
            "schema": {
              "type": "object"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "OrderCart": {
      "description": "cart Id",
      "required": [
        "id"
      ],
      "properties": {
        "id": {
          "description": "cart Id",
          "type": "string"
        }
      }
    },
    "UpdateOrderCart": {
      "description": "cart Order",
      "required": [
        "id",
        "quantity"
      ],
      "properties": {
        "id": {
          "description": "cart Id",
          "type": "string"
        },
        "quantity": {
          "description": "cart Id",
          "type": "number"
        }
      }
    },
...
  }
}